Work in progress toward style assets
This now adds a Makefile to retrieve and place some theme assets from
the style guide repo. This is the approach used by the Rails theme gem,
but it feels a bit awkward here at the moment. Given what we are doing
with PHP and SASS, it feels like we'll need yet another GH Actions file
to make sure that the developer has run `make update` prior to pushing.

Under this approach, we own everything in the body tag - so while we
are copying the layout.phtml file, we then end up replacing the (very
minimal) contents of the body tag in the next step.

The partials in view/mitlib-style have some Rails-isms that need to be
replaced. The next commit shows an example of what that looks like.

The favicon referenc in layout.phtml is static, and needs to be
updated to something populated by PHP.

On the plus side, the files in any mitlib-style folder are now placed
by the Makefile (as with the libraries-main.min.css file) - so really
this PR could be:
* .gitignore
* Makefile
* config/theme.ini (version change)
* view/layout/layout.phtml

That's a pretty small PR, with the rest the output of the Makefile.

(This commit should probably be updated to be the result of those
changes, and then one that's made by the Makefile)