Decision Record 4
Why these changes are being introduced:
This is the first pass at a decision record for how we will implement
the automated deployment of the theme to the hosted Omeka S servers. I
want the decision reviewed before beginning any work in the repository.

How this addresses that need:
* Add the decision to the ADR folder

Side effects of this change:
None.

Relevant ticket(s):
* See the https://mitlibraries.atlassian.net/l/cp/yNV9fsVz document in
Confluence