Adds a feature for color variations to theme
** Why are these changes being introduced:

We want to allow site builders to make minimal customizations to how
their theme displays, using colors inspired by the Pentagram branding
but still with sufficient color contrasts for the areas that receive
the highlight color.

This is a feature used elsewhere in the Omeka community, for example by
Union college as well as by the Foundation theme.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/post-65

** How does this address that need:

This adds a field to the theme settings form, providing a dropdown of
color choices. The default is a black variant, without a highlight color
(and a blue hover state for text links). Other variants include:

* MIT colors (red and silver-gray)
* Blue
* Burnt orange
* Magenta
* Red

Each variant has its own stylesheet, which defines colors in semantic
variables. Those variables are then used in SCSS modules, which will
hopefully make maintenance of these variants easier to manage over time.

Adding new variants should be straightforward, requiring only a new
intermediate stylesheet (see asset/scss/red.scss for an example).

** Document any side effects to this change:

There are some changes to existing stylesheets in order to accommodate
this feature:

* We unset the border-radius on the navigation menu, and specify a
  white text color on highlighted menu options.
* We are fixing the line-height on blockquote elements.
* Floated media-embed blocks now have a slight amount of padding, so
  that the drop shadow doesn't get clipped.
* The minimum height of the banner element goes to 12rem, so that it
  should always be visible.
* The top part of the masthead element goes from gray to black.