Theme Deployment Automation Configuration
Why these changes are being introduced:
We intend to fully automate the theme deployment for the Omeka S
servers. This includes workflows for GitHub actions along with full
documentation of the decision and configuration.

How this addresses that need:
* Create a GitHub Actions workflow for our stage environment that will
run on any merge to the `main` branch that includes files that affect
the theme
* Create a GitHub Actions workflow for our prod environment that will
run on any published release
* Create a `deploy` command in the Makefile that handles the commands
to copy the theme files to the correct location on the Omeka host
server
* Update ADR #4 to switch the status from "proposed" to "accepted"
* Create a reference doc that describes the deployment automation in
detail
* Update the main README with information about the deployment process

Side effects of this change:
None until this is merged to `main`. At that point, future merges and
releases will deploy automatically.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/POST-60