Address PR Comments
Why these changes are being introduced:
Still working through exactly what we want to track in CODEOWNERS and
also addressing the desire to be able to deploy interim theme updates
for stakeholder review BEFORE merging a PR to the main branch. Also
working through some tweaks to the GHA workflows that will best serve
the developers.

How this addresses that need:
* Fix the Makefile deploy commands
* Add the Makefile to the CODEOWNERS file
* Update the title for ADR#4
* Create a new workflow_dispatch action to all devs to manually trigger
a deploy on the staging server for stakeholder review
* Leverage the "job output" feature of GitHub actions to echo the value
of the "version" from the config/theme.ini file

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/POST-60