Fix a11y bug in skip link
There was a bug in our layout template, where the skip link at the top
of the template was pointing to #content, but the actual content div
had an id value of content-main.

This updates the skip link to point to the correct id value (I can't
rule out that the id value itself is used in our styles somewhere, so
it seemed easier to update the skip link).