Implement custom styles and template changes
** Why are these changes being introduced:

* We need to write some custom CSS styles to make page records appear
  in a way consistent with our design system that also matches what we
  want our theme to provide for Omeka specifically.

* In order to make these changes happen, we also need to add an image
  size name as a class value to the item showcase template.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/post-84

** How does this address that need:

This sets up a system of SCSS modules in the asset/scss folder, which
roughly correspond to one file for every block type used on pages. The
exception is that we're also adding a "resource" module that will take
effect across multiple blocks.

These modules get compiled into the overall stylesheet, similar to all
the other modules in the theme. There is still one overall stylesheet
for the theme.

Almost all of the style changes we're making deal with how images are
presented. There are links in the Jira ticket to pages that demonstrate
how all this takes effect.

Non-image changes add some spacing under map blocks, tweak the way a
list of sites appears, and raises the default font size and line height
for the entire theme to make things a bit more readable.

** Document any side effects to this change:

* This commit also bumps the theme version to 0.5