Copy-paste templates that need customized
The work of these tickets require us to make changes to some view
templates that Omeka uses internally.

For this to happen, we need to first copy-paste those templates from
their native location to the same location inside the theme structure.

The copied templates are:

* view/common/linked-resources.phtml
* view/common/resource-page-block-layout/values.phtml
* view/omeka/site/index/search.phtml
* view/omeka/site/item/browse.phtml
* view/omeka/site/item/show.phtml

The side effect of this change is that we now own these templates. It
would be good to keep tabs on how their original templates change -
although this is not required. The more fundamental risk is that Omeka
retires any of the internal methods or classes which the templates use.