Customize advanced search templates
** Why are these changes being introduced:

* In order to make the advanced search form render the way that we want,
  some changes are necessary to the templates which control the markup.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/post-114

** How does this address that need:

* The big change is that, now that the advanced-search.phtml template is
in a theme, we no longer need the conditionals which influence how the
search form appears within the admin interface. So we remove them to
make the template easier to follow. These are all the checks for
$this->status()->isAdminRequest().

We also remove the search block to allow searching by item set, because
the sites we will build will only have one item set per site - making
the control meaningless.

We are also not searching for media or for item sets, so those paths
are removed. We don't take out the entire switch because I can't ensure
that it is _never_ needed, though.

* The fulltext field is converted from a single line input to a text
area, which then has the class values our style library expects.

* The search page (omeka/site/item/search.phtml) also has the expected
classes added to the submit button.

** Document any side effects to this change:

* I'm not sure whether the switch in advanced-search.phtml should be
removed, so that might be a code smell. For now, though, I'm leaving
it in due to caution. This might be considered a side effect?