Theme setting to suppress links to advanced search
** Why are these changes being introduced:

* The advanced search form will likely not be needed when the platform
launches, but eventually we expect that it will be desired and ready to
use.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/post-114

** How does this address that need:

* This defines a theme setting to enable and disable the link to the
advanced search form. It also adds a check on the item browse template,
which respects this setting.
* The setting is placed in the "browse settings" area of the theme
settings form.

** Document any side effects to this change:

* There is another location where the advanced search link can appear,
on the item set browse - but we have previously decided that we won't
be using that browse - sites will only ever have one item set.

We have already made other customizations to the item browse that
might need to be ported to the item set browse - so this is one more
such difference. To be future-proof, we might want to extend this
setting (and our other customizations) to the item set browse - but
that does not need to happen now. I just note it here, as a reminder.