Start hacking SmartMenus
Having copy-pasted the un-minimified version of the SmartMenus in the
previous commit, now we switch to using that in the theme, and start
making changes.

First up is to change the tag used for the submenu indicator, from a
span to a button.

We also experiment with the invocation of SmartMenus, using the
undocumented ability to use values for subIndicatorsPos other than
"prepend" and "append". Specifically, we use "after" to get the
sub-indicator to be inserted outside of the menu link.

This breaks the ability to interact with the indicator, so further
changes are going to be needed. Additionally, some CSS is going to
need updating because the link currently displays as block - forcing
the sub-indicator to slide below.