Refactor navigation templates to allow custom markup
Our previous implementation of navigation templates used the default set of
markup generated by Omeka (and Laminas beneath that). However, the menu
library this implements makes some different assumptions about the markup
that will be present when it starts.

As a result, this replaces the previous minimal menu templates with a pair
of custom rendering templates. The pattern followed here was borrowed from
the Foundation theme, which uses very similar templates to build its menus.

Some of the logic in the Foundation templates has not been replicated here,
partly because we don't have some of that theme's features, and partly
because some conditional logic in those templates doesn't seem to correspond
to anything that I can identify.