Fix Theme Deployment Logic
Why these changes are being introduced:
The previous version of the deploy-to-prod workflow had a logic mistake.
There was a step in the job that checked for the branch, but the exit
code implemented in that step just exited the step and did not
actually exit the job! So, tags on branches other than `main` would
actually trigger the prod-deploy (oops). The fix here is based on this
StackOverflow conversation:
https://stackoverflow.com/questions/71535871/trigger-github-action-only-when-new-tags-on-master-branch

How this addresses that need:
* Move the branch check logic into an embedded `if:` line in the deploy
step itself, so that the step will only run if the branch is `main`.

Side effects of this change:
None.