Testing Solution for GHA Problems
Why these changes are being introduced:
The Omeka hosts are running Amazon Linux 2, which is build on CentOS 7
which is no longer on the GHA compatibility list. The shift to node20
(from node16) for all GHA runs requires newer versions of the glibc
libraries that are not available for AL2. This means that our automated
deployment workflows that use the Omeka hosts as our self-hosted GHA
runners will now fail to run because of the glibc dependency.

How this addresses that need:
* This is an attempt to override the requirement for a newer version of
glibc

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1327