Replace actions/checkout Functionality
Why these changes are being introduced:
The hosts running the GitHub actions are running Amazon Linux 2 which
is no longer on the supported host OS list for self-hosted GitHub
runners. Since the actions-runner on the hosts is configured to
auto-update, it not longer has node16 embedded and node20 cannot run on
AL2 (because the host OS does not have an up-to-date-enough version of
the glibc library).

The only thing in the workflow that requires node20 is the
actions/checkout command. So, instead of using actions/checkout, we just
run the very short list of git commands that we need to checkout the
correct commit to run the simple deploy command.

We update the manual workflow, the staging workflow (for pushes to the
main branch) and the production workflow (for tagged releases on the
main branch).

How this addresses that need:
* Remove the actions/checkout@v3 step in all three workflows
* Create new steps for the necessary git commands to clone the repo and
checkout the correct commit

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1327