Bug Fix: Update Workflows
Why these changes are being introduced:
The last fix was an incomplete fix since I assumed that the runner
environment was transient. But, with a self-hosted runner, the
file/folder structure remains in place after each run. That resulted
in the `git clone` step failing the second time the action ran since
there were already files in the target directory.

How this addresses that need:
* Add an `rm` step at the start of each of the three deployment
workflows to remove the "repo" directory before the `git clone` command
runs to the pull the latest commit into the folder

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1327