Remove extraneous descriptive text in _layouts.scss; stylelint version in package.json; package-lock.json integrity sha512- value update(?)
Remove verbose reference to the date of my changes to the file in _layouts.scss

We don't fully know what the changes to pakcage-lock.json and package.json mean. Changes to integrity sha512-  and stylelint version package.json. Matt B. didn't think whatever warranted the commit was significant but to go ahead and commit anyway. We didn't ourselves touch the files.